Let the extractor accept an initial HTML source.
If provided with an initial HTML content, the extractor will not load
the page and use it as its content. It will then processed normally.

The extraction worker will not attempt to follow links (main page or
page sequence) when given an initial HTML.

The initial HTML can be sent to the API with a multipart/form-data
payload.