internal/db improvements. Preparation for multidb support.
The connection configuration does not need a "driver" value, it's given
by the scheme already, so let's just use it.

Added a migration system. Each applied migration is stored in the DB
with its number, name and date. Migrations are taken from embed files
depending on the database driver in use.

Added some utility functions in internal/db to deal with the
different ways databases like to do things.