observe mutations on elements in a turbo-refresh.
Instead of checking the presence of a target, we'll check the
presence of a given selector in the controller's element children.

We do so when the controller connects and then on any mutation in the
controller. As long as the selector is present, a refresh is perform
with a timeout.