Added ErrorPages middleware.
This middleware wrap the http.ResponseWriter to a new one that will
intercept status codes and, later, replace the content sent to the
user with html, json or plain text, depending on the request Accept
header.

The response defaults to text/plain is no accept header is received and
to text/html if an Accept header is present but does not yield anything
we can process. A browser always sends an Accept header, so defaulting
to text/html is safer in this case.