Rule engine using JS files.
This might sound a bit overkill to fire up some JS engine but that's
the easiest way to provide simple and fast rule engine with two way
bindings.

Rules have a tiny API that can help grab a JSON file for a given
resource and override some metadata on the current drop being processed.

Later on, we can provide a way to have a userscripts configurable folder
to install specific rules for a given installation.