Keep focus on label input field when submitting a label.
The turbo-stream was sending the full label form on submit, hence
replacing the current form and then losing the focus on the input
field. Now we only return the label list.

Since, the input field stays in place, we have to empty it after
the data has been submitted. That's done on turbo:submit-end for any
element in the target with data-turbo-empty-submit-end.